ai-engine/requirement.txt
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib
seaborn
